Fix non-trivial warnings hidden by -isystem
These warnings were hidden by -isystem.

Bug: 31751828
Test: m -j checkbuild
Merged-In: Ibb1aa07a6827fcbd29fc15291fcd90280c0f608d
(cherry picked from commit 6f51c153ea5e4187a313e47ac8bad9a1a999d4e1)

+++ media/libstagefright/wifi-display/MediaSender.cpp
@@ -420,7 +420,7 @@ status_t MediaSender::packetizeAccessUnit(
             CHECK(accessUnit->meta()->findInt32("rangeLength", &rangeLength));
             CHECK(accessUnit->meta()->findMessage("notify", &notify)
                     && notify != NULL);
-            CHECK_GE(accessUnit->size(), rangeLength);
+            CHECK_GE((int32_t)accessUnit->size(), rangeLength);
 
             sp<GraphicBuffer> grbuf(new GraphicBuffer(
                     rangeOffset + rangeLength /* width */, 1 /* height */,
